Extract ApplicationChoice-related actions from CourseChoicesController
Move the destroy, review, and complete actions. The behaviours
here relate largely to the ApplicationChoices created for the
candidate's application, so move them to an
ApplicationChoicesController.
