+++ app/views/candidate_interface/application_choices/_guidance.html.erb
