+++ app/views/candidate_interface/application_choices/confirm_destroy.html.erb
