+++ app/views/candidate_interface/application_choices/index.html.erb
